chore: initial commit; add files
Most of the initial development was done in the repo that houses my
personal website. Initially, I thought this blog generator would be a
Python script. However, the code got to the point where I wanted to
split it into multiple files. Also, someone asked me if they could use
it. Since I might not be the only person using this blog generator, I
thought it would make sense to give it its own repo and publish it to
PyPI.